Expose function in the server to read config file
This is entirely refactored from code that existed already in
cmd/gobgpd. It is useful for apps that want to embed gobgp and still
benefit from configuring the server through the regular config file
format.

This commit decomposes the config file processing code from gobgpd
into smaller methods which are easier to understand and makes the flow
clearer.