Avoid mutex to set idle hold time after restart
This uses a non-blocking select on a channel of depth 1. It also uses
a non-blocking send. When reset occurs, the non-blocking select
ensures that there is exactly one value in the channel the next time
the code enters the idle state.

When entering the idle state, the non-blocking receive ensures that
the channel is left empty. If it read a value, then it is used as the
idle timeout. Otherwise, it uses the default.

The effect is the same and avoids mutex sychronization making the code
easier to understand.