Fix tools test
This changes the test to no longer reference $GOPATH. Instead, it
removes `github.com/osrg/gobgp/` fore calling grep from the package
name before calling grep so that it greps a location relative to
wherever the script was located when it was called. This should work
regardless of $GOPATH.

The problem was that in a Travis CI run, the repository is not cloned
to the correct spot under $GOPATH and the grep fails because it is
looking in the wrong place. The repository is cloned to
`/home/travis/build/osrg/gobgp` and GOPATH is set to
`/home/travis/gopath` during a Travis run which don't match up.