fix duplicated local path id bug
Fix a bug that the same path id is assigned to two paths. The bug
happens in the following way:

1. a new path is assigned to a local path id.
2. an import policy dropping the path from the master rib is added and
execute softreset in.
3. the path still exists in the adj but doesn't in the master. the
path id was freed (marked as unused).
5. a new path with the same prefix comes from another peer. The same
id is assigned to the new path.
6 deleted the policy and execute softreset in.
7 there are two paths in the master with the same path id.

This path guarantees that only when a path is removed in the adj, the
id for path is freed.

Note that this doesn't fatten Path strcuture, which should be avoided
for any reason.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>