tests: fix failing tests when using more recent versions of protobuf
When regenerating the `*.pb.go` file, we get many tests failing
because testify is using `reflect.DeepEqual()` under the hood. There
is no option to have a custom comparison function and currently, the
only way around is to use `proto.Equal()` to compare protobuf
messages. The downside of using `proto.Equal()` is that we loose the
expected/actual display in case of a mismatch.

When comparing family, we compare `Api`/`Sapi` in separate asserts
instead of using `proto.Equal()`.

Fix: #1952