pkg/packet/bgp: initial support for BGP VPLS
This includes BGP VPLS NLRI and extended community. There is no CLI/API
support yet. Also, VPLS AFI/SAFI is shared with RFC 6074 which says:

> Both BGP-AD and VPLS-BGP [RFC4761] use the same AFI/SAFI.  In order
> for both BGP-AD and VPLS-BGP to co-exist, the NLRI length must be
> used as a demultiplexer.
>
> The BGP-AD NLRI has an NLRI length of 12 bytes, containing only an
> 8-byte RD and a 4-byte VSI-ID.  VPLS-BGP [RFC4761] uses a 17-byte
> NLRI length.  Therefore, implementations of BGP-AD must ignore NLRI
> that are greater than 12 bytes.

This is not handled in this PR.

The generic transitive experimental extended community types are shared
with FlowSpec. The parsing is extended to handle VPLS as well, but not
other L2 encapsulations.

Fix #2590