server: disable tcpkeepalive
The default value for tcpkeepalive in the standard net
package is 15 seconds, while the default value for BGP
keepalive is 30 seconds. If the network is not good,
the TCP connection is disconnected before BGP can send
keepalive messages.

Therefore, BGP keepalive is effectively invalid.

The solution is that since we have BGP keepalive, we
don't need to open tcp keepalive.

Signed-off-by: Chen Yahui <goodluckwillcomesoon@gmail.com>