cmd/gobgp: fix parseEvpnMacAdvArgs
default-gateway not being declared a reserved keyword led to evpn
macadv commands ending as "encap vxlan default-gateway" to fail parsing.

In that case, default-gateway was bundled with the previous argument.

To fix that, set default-gateway as a reserved keyword of paramFlag type.

While there, change the way we detect its existence later on, given that it now
exists in our reserved map m, if set.