test/evpn: add multi-vrf mac-mobility test
This test ensures that a MAC can be announced by different peers as long
as it is in different MAC-VRFs, not applying MAC-mobility algorithms.

It does so by adding two additional VRFs for the new route to land in,
and by cleaning up a bit the RDs and RTs used by the existing tests.