make config parsing strict when starting gobgpd
The default policy for gobgp if config is invalid is to start with
default policy ACCEPT. This is harmful for our Route Reflectors
setup in which we can get spurous routes after unlucky release.

This commit adds --config-strict option that makes gobgp do the
fatal exit if config is invalid. If option is not specified,
gobgpd will still log warning and continue working.

It also introduces special facilty logging field similar to the
one in syslog. This allows to implement said backward compatibility
with Fatal() calls with facility "config" being downgraded to Warn().

Reload behavior is not changed as it is not as harmful (policies
are retained in such case).