Add FSM loop timings
As the issue #2854 demonstrates, Main FSM loop performance is crucial
for overall GoBGP performance. While we're far from hitting those limits yet,
we want to be sure that we have enough capacity in FSM loop.

To achieve that, we add FSM loop timing histograms, which are showing:
    - timing_sec - amount of time handling each message took
    - wait_sec - amount of time each message spent in a channel before
      it was received by the main FSM loop.
The latter (wait times) allows to detect situation when FSM loop performance
is inadequate, while the former (timings) allows to understand which
component causes such delays.

For the sake of simplicity, further classification of message types, such
as RouterID of the peer or name of gRPC operation is not made.