Add BufferChannel for fsm incoming/outgoing channels.
Added the ability to select a type of fsm incoming/outgoing channels:
1. InfiniteChannel (old behavior, default)
2. BufferChannel

BufferChannel allows to accumulate and combine paths in a separate thread, in the case when the reading side (for example, BgpServer.Serve) is busy.
A map with NLRI key is also used to combine multiple changes of one path into a single update.