fsm: Speed-up initial connection time
In the existing implementation, the initial connection
from GoBGP side was not attempted until the timer
with a random value from the interval [5, 10) seconds expired.

This change makes the initial connection attempt immediate,
and lowers the minConnectRetryInterval to 1s. This helps to
reduce convergence times, which is helpful especially
in the datacenter deployment scenarios.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>