Fix invalid nexthop advertisement with nexthop-unchanged
When we apply nexthop-unchanged action to the locally-originated route
with unspecified nexthop (0.0.0.0 or ::), we advertise the path with
unspecified nexthop. This may ends up with the session shutdown with
Invalid Update Message Notification (only tested with FRR 10.2.1).

Fix this issue by setting table.PolicyOptions.OldNextHop to the local
peering address when the path is locally-originated and has an
unspecified nexthop.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>