feat: insert new paths into destination's known path list in sorted mannger
The current scheme for sorting a destination struct's known path list
relies on golang's SliceStable, which at scale, is an O(n * log n)
operation. However, if the sorting is done at the same of adding a new
path, then we can assume rely on an O(n) operation instead. This assumes
that `Calculate` is the only code block where a unique destination for
the global RIB is being manipulated.