feat: Allow graceful restart in StopBgpRequest
Before this change, it was not possible to stop BgpServer
while allowing Graceful Restart to kick in on the remote peers,
as we send a NOTIFICATION mesage which terminates GR to each peer
upon Stop(). This change allows to skip sending the NOTIFICATION
messages to GR-enabled peers if the allow_graceful_restart flag
is set to true. Keeps the existing behavior if the flag is not set.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>