fix: wait for out msg to be sent
issue raised by the TestFSMHandlerEstablish_HoldTimerExpired test:
when the hold timer expires, we push a message to the outgoing
infinite channel that rely on another goroutine to push it to the
other side of the infinite 'channel'.
However, since we cancel the 'sendMessageLoop' context when returning
from the 'established' function, a race occurs between the
'sendMessageLoop' function ending and the message being pushed.
to fix this, using a native go channel solve the issue.