Reduce randomness in active connect retry delay
In the current implementation, the wait time before the next active
connection attempt can be up to twice the user-specified connect-retry
interval. For example, if connect-retry is set to 10 seconds, the
actual wait time can range from 10 to 20 seconds due to an added
random delay.

RFC 4271 says:

  To minimize the likelihood that the distribution of BGP messages by a
   given BGP speaker will contain peaks, jitter SHOULD be applied to the
   timers associated with MinASOriginationIntervalTimer, KeepaliveTimer,
   MinRouteAdvertisementIntervalTimer, and ConnectRetryTimer. A given
   BGP speaker MAY apply the same jitter to each of these quantities,
   regardless of the destinations to which the updates are being sent;
   that is, jitter need not be configured on a per-peer basis.

   The suggested default amount of jitter SHALL be determined by
   multiplying the base value of the appropriate timer by a random
   factor, which is uniformly distributed in the range from 0.75 to 1.0.
   A new random value SHOULD be picked each time the timer is set. The
   range of the jitter's random value MAY be configurable.

Following the RFC's suggestion, the interval is determined by
multiplying the configured connect-retry value by a random number in
the range of [0.75, 1.0). So this makes the actual interval closer to
the configured interval.

Although the RFC does not state whether to apply jitter before the
initial attempt, we will use a value calculated by multiplying 2000
milliseconds by a random number in the range [0.75, 1.0). Note that
the current implementation uses a value in the range [1000, 2000).

The 2000 milliseconds was chosen arbitrarily, based on the assumption
that most GoBGP use cases are within data centers, where long timeouts
are undesirable.

This change is expected to reduce the likelihood of peer connection
timeouts in CI environments.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>