pkg/packet/bgp: use netip for IPAddrPrefix and IPv6AddrPrefix
Replace uses of net.IP with netip.Prefix for IPAddrPrefix and
IPv6AddrPrefix.

When GoBGP was started, the net/netip package did not exist and net.IP
was the standard type for representing IP addresses. However, with the
introduction of netip.Addr in Go 1.18, we now have a more efficient
and type-safe alternative.

This patch replaces uses of net.IP with netip.Prefix type implemented
by using netip.Addr for IPAddrPrefix and IPv6AddrPrefix.

We should use netip for IpAddrPrefixDefault however it means that this
patch becomes much larger since we also update all labeled addrprefix
types. So IpAddrPrefixDefaultNetip is added temporarily. There will be
some temporary code duplication, but I plan to gradually proceed with
the replacement.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>