pkg/table: make AddrPrefixOnlyCompare private
Move AddrPrefixOnlyCompare from bgp to table package; make it private
function.

Public API interface is written in stone. It cannot be changed without
the major version updated. Currently, AddrPrefixOnlyCompare is used
only by table package so let's keep it in the table package. We can
always make it public in the future when needed.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>