server/fsm: move loopback check to fsmHandler
We don't need to check whether the peer address is a loopback on every
received UPDATE message.

This was added for testing multiple GoBGP instances on localhost

16a9c953 ("Allow testing multiple GoBGP instances on localhost.").

We allow only v4 here because In IPv4, we can pick two different
addresses from 127.0.0.0/8, have two peers bind to those addresses,
and establish a TCP connection over the loopback device. In IPv6,
however, the loopback address is limited to ::1, so with the current
implementation (which distinguishes peers by IP address) this likely
would not work (I've not tried).

This is preparation for removing peerInfo in fsm.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>