table/path: Add bgp.Family to Path
Add bgp.Family to the Path structure in preparation for removing AFI()
and SAFI() in bgp.AddrPrefixInterface interface.

The Path structure has unused space, so its size does not change
(verified on a 64-bit architecture).

In the long term, the family info should be placed in PeerInfo, but
since that would require significant changes. The Path size will not
increase, so we chose a simpler implementation.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>