config: use netip for some OpenConfig types
- netip.Addr for inet:ip-address, inet:ipv4-address, and
  bgp-types:rr-cluster-id-type

- netip.Prefix for inet:ip-prefix

Add a workaround for union type for bgp:local-address to
bgpyang2golang.py. Go doesn't support union type so pyang has to
choose one type. But pyang doesn't know which is appropriate so
bgpyang2golang needs to select an appropriate type for a member.

Also there are some changes for GoBGP specific types:

- netip.Addr for BGP remote ID instead of string
- netip.Addr for local addresses to listen instead of string
- netip.Prefix for Dynamic peer config instead of string

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>