api: remove ApplyPolicy from Global
The ApplyPolicy field in the API Global structure was added as an
equivalent to config.Global.ApplyPolicy used in file-based
configuration. However, unlike file-based configuration, API-based
configuration is designed to be set through multiple policy-related
APIs such as AddPolicy, making ApplyPolicy inappropriate and unused in
the API context. Therefore, it should be removed.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>